#12151: Move max_pool2d to generic
