+++ ttnn/cpp/ttnn/operations/pool/generic/device/max_pool2d_multi_core_program_factory.cpp
@@ -320,15 +320,15 @@ MaxPool2D::MultiCore::cached_program_t max_pool_2d_multi_core_sharded_with_halo_
     std::string reader_kernel_fname;
     if (is_large_kernel) {
         reader_kernel_fname =
-            "ttnn/cpp/ttnn/operations/pool/maxpool/device/kernels/dataflow/"
+            "ttnn/cpp/ttnn/operations/pool/generic/device/kernels/dataflow/"
             "reader_max_pool_2d_multi_core_sharded_with_halo_large_kernel_v2.cpp";
     } else if (is_wide_reduction) {
         reader_kernel_fname =
-            "ttnn/cpp/ttnn/operations/pool/maxpool/device/kernels/dataflow/"
+            "ttnn/cpp/ttnn/operations/pool/generic/device/kernels/dataflow/"
             "reader_max_pool_2d_multi_core_sharded_with_halo_wide.cpp";
     } else {
         reader_kernel_fname =
-            "ttnn/cpp/ttnn/operations/pool/maxpool/device/kernels/dataflow/"
+            "ttnn/cpp/ttnn/operations/pool/generic/device/kernels/dataflow/"
             "reader_max_pool_2d_multi_core_sharded_with_halo_v2.cpp";
     }
 
@@ -374,10 +374,10 @@ MaxPool2D::MultiCore::cached_program_t max_pool_2d_multi_core_sharded_with_halo_
     std::string compute_kernel_fname;
     if (is_large_kernel) {
         compute_kernel_fname =
-            "ttnn/cpp/ttnn/operations/pool/maxpool/device/kernels/compute/max_pool_multi_core_large_kernel.cpp";
+            "ttnn/cpp/ttnn/operations/pool/generic/device/kernels/compute/max_pool_multi_core_large_kernel.cpp";
     } else {
         // both regular and wide reductions
-        compute_kernel_fname = "ttnn/cpp/ttnn/operations/pool/maxpool/device/kernels/compute/max_pool_multi_core.cpp";
+        compute_kernel_fname = "ttnn/cpp/ttnn/operations/pool/generic/device/kernels/compute/max_pool_multi_core.cpp";
     }
 
     auto compute_kernel = CreateKernel(program, compute_kernel_fname, core_range, compute_config);
